refactor(v0.71.0): reports.py 1156L → 12-module 패키지로 분할 (Tier 3 #3) (#897)
Tier 3 God Object 분할 시리즈 세 번째. ReportGenerator 클래스 +
33개 module-level 포매터 함수가 한 파일에 묶여 있던 1156 LOC monolith를
보고서 관심사별 12개 모듈로 분리.

새 패키지 구조 — core/skills/reports/:
  __init__.py            46L  12개 심볼 re-export (외부 2곳 import 변경 0)
  models.py             107L  ReportFormat/ReportTemplate enums + 템플릿 로더 +
                              tier 헬퍼 + gauge 기하학
  scoring.py            150L  subscores/synthesis/analyses html+md
  evaluators.py          86L  eval 필드 추출 + table 포매터
  psm.py                158L  PSM + scoring breakdown
  biasbuster.py          72L
  signals.py             94L
  analyst_reasoning.py   71L
  cross_llm.py           60L
  rights_risk.py         62L
  decision_tree.py       75L
  generator.py          336L  ReportGenerator 클래스 (의도적 단일 보존)

templates/ 디렉토리는 git mv로 core/skills/reports/templates/ 이동 —
히스토리 보존, Path(__file__).parent / "templates" 의미 유지.

공개 API 변경 0건. 외부 import 2곳 (core/cli/report_renderer.py +
tests/test_reports.py) 수정 불필요 — 패키지 __init__.py 가
ReportFormat/ReportGenerator/ReportTemplate + 8개 _format_* 모두 re-export.

.gitignore 조정: 78번 줄 'reports/' 패턴이 새 패키지 경로를 가리는 문제
회피용 negation 추가 (lines 79-83). __pycache__는 정확히 제외하도록
범위 명시.

Quality Gates:
- ruff/ruff-format/mypy 0 errors (279 source files, +11)
- pytest 4344 passed, 1 skipped, 24 deselected — 변동 없음
- 보고서 + E2E 75개 테스트 통과
- E2E Cowboy Bebop A (68.4) 변동 없음
- 최대 단일 파일 1156 → 336 LOC (-71%)

남은 Tier 3 후보 6개 — commands.py(2441), cli/__init__.py(1889),
agent/loop.py(1754), agentic_ui.py(1160), tool_executor.py(1047),
llm/router.py(1046).
