+++ core/skills/reports/templates/report_detailed.md
